# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DreamCraft API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/marks": {
            "post": {
                "tags": [
                    "Mark"
                ],
                "summary": "Create mark",
                "operationId": "1be556f12770f7bfbc5b1d0d54bfd839",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "mark": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "category_id": 1,
                                    "mark": 5
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create mark",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "data",
                                                    "example": "2024-04-27T21:27:04.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "data",
                                                    "example": "2024-04-27T21:27:04.000000Z"
                                                },
                                                "mark": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "The range of evaluation has been breached | Duplicate mark",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Evaluation": {
                                        "summary": "Evaluation",
                                        "value": {
                                            "message": "You cannot give a rating lower than 1"
                                        }
                                    },
                                    "Duplicate": {
                                        "summary": "Duplicate",
                                        "value": {
                                            "message": "You already rated this category"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "data": "Category not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "message": "The mark field must be an integer.",
                                            "errors": {
                                                "mark": "The mark field must be an integer."
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "message": "Something wrong. Try later."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}